Rock Paper Scissors
Rock, paper, scissors, also know as roshambo, is a simple child's game
that is frequently used to settle disputes. In the game, a rock breaks
the scissors, the scissors cut the paper, and the paper covers the rock.
Each option is equally likely to prevail over another. If the players
choose the same object a draw is declared and the game is repeated until
someone prevails. In this computerized version the human player competes
against the computer which chooses a rock, paper, or scissors randomly.
The game proceeds until the human player quits the game or until a
predetermined score is reached.